test: refactor restricted field tests
