+++ packages/amplify-graphql-api-construct-tests/src/__tests__/restricted-field-auth/graphql-schemas/gen1-ddb-only/API.ts
