+++ packages/amplify-graphql-api-construct-tests/src/__tests__/restricted-field-auth/graphql-schemas/gen2/schema-related.graphql
@@ -1,19 +1,7 @@
-input AMPLIFY {
-  engine: String = "mysql"
-}
-
-type Primary @model @auth(rules: [{ allow: public, operations: [read] }, { allow: owner }]) {
-  id: String! @primaryKey
-  secret: String @auth(rules: [{ allow: owner }])
-  owner: String @index(name: "primary_owner")
-  relatedMany: [RelatedMany] @hasMany(references: ["primaryId"])
-  relatedOne: RelatedOne @hasOne(references: ["primaryId"])
-}
-
 type RelatedMany @model @auth(rules: [{ allow: public, operations: [read] }, { allow: owner }]) {
   id: String! @primaryKey
   secret: String @auth(rules: [{ allow: owner }])
-  owner: String @index(name: "related_many_owner")
+  owner: String
   primaryId: String
   primary: Primary @belongsTo(references: ["primaryId"])
 }
@@ -21,7 +9,7 @@ type RelatedMany @model @auth(rules: [{ allow: public, operations: [read] }, { a
 type RelatedOne @model @auth(rules: [{ allow: public, operations: [read] }, { allow: owner }]) {
   id: String! @primaryKey
   secret: String @auth(rules: [{ allow: owner }])
-  owner: String @index(name: "related_one_owner")
+  owner: String
   primaryId: String
   primary: Primary @belongsTo(references: ["primaryId"])
 }
